[PATCH] M68k update (part 8)
M68k bitops updates
  - use bitmap_member() for bitops data declaration
  - Make the m68k bitops really operate on unsigned long
  - Add fls()

+++ include/linux/generic_serial.h
@@ -45,7 +45,7 @@ struct gs_port {
   int                     count;
   int                     blocked_open;
   struct tty_struct       *tty;
-  int                     event;
+  unsigned long           event;
   unsigned short          closing_wait;
   int                     close_delay;
   struct real_driver      *rd;
